+++ src/main/java/com/cidenet/kardexapp/commons/domains/generic/OutDTO.java
@@ -17,8 +17,8 @@ public class OutDTO {
     private Timestamp lastModified;
     private Integer kardexId;
     private String observation;
-    private Integer unitValue;
-    private Integer totalValue;
+    private Double unitValue;
+    private Double totalValue;
     private Timestamp date;
 }
